Changes app initialization for better cluster support.
