--- src/app.ts
@@ -1,9 +1,0 @@
-'use strict';
-
-import { start } from './start';
-
-start()
-    .catch((err) => {
-        console.error(`Error starting gateway: ${err.message}`);
-        process.exit(-1);
-    });
