+++ src/index.ts
@@ -1,0 +1,8 @@
+'use strict';
+
+import './command-line';
+import { Application } from './application';
+
+const app = new Application();
+
+app.start();
